requirements.txt
agent-sandbox>=0.0.19
dotenv>=0.9.9
fastapi>=0.115.0
httpx>=0.28.0
kubernetes>=30.0.0
langchain>=1.2.3
langchain-anthropic>=1.3.4
langchain-deepseek>=1.0.1
langchain-mcp-adapters>=0.1.0
langchain-openai>=1.1.7
langgraph>=1.0.6
langgraph-api>=0.7.70
langgraph-cli>=0.4.14
langgraph-runtime-inmem>=0.22.1
markdownify>=1.2.2
markitdown[all,xlsx]>=0.0.1a2
pydantic>=2.12.5
python-multipart>=0.0.20
pyyaml>=6.0.3
readabilipy>=0.3.0
sse-starlette>=2.1.0
tavily-python>=0.7.17
firecrawl-py>=1.15.0
tiktoken>=0.8.0
uvicorn[standard]>=0.34.0
ddgs>=9.10.0
duckdb>=1.4.4
langchain-google-genai>=4.2.1
langgraph-checkpoint-sqlite>=3.0.3
langgraph-checkpoint-mysql[aiomysql]>=1.0.0
langgraph-checkpoint-mysql[asyncmy]>=2.0.4
lark-oapi>=1.4.0
slack-sdk>=3.33.0
python-telegram-bot>=21.0
langgraph-sdk>=0.1.51
markdown-to-mrkdwn>=0.3.1
langgraph-checkpoint-mysql>=3.0.0
pymysql>=1.1.2
asyncmy>=0.2.10
forbiddenfruit>=0.1.4
python-docx
docxtpl
doc2docx
requests
rapidfuzz>=3.0.0
tenacity>=8.2.0